create structure.js file with array, container and count
